refactor: reorganize module structure and enhance I/O functionality
- Moved I/O related files from `src/pgn/` to `src/io/`, improving module clarity and separation of concerns.
- Updated references throughout the codebase to reflect the new module paths, ensuring consistency in imports.
- Introduced subpath exports for `io`, `pgn`, and `replay` in package.json, enhancing usability for consumers of the library.
- Revised documentation in AGENTS.md, CHANGELOG.md, and README.md to reflect the new module layout and clarify usage.
- Enhanced tests to validate the functionality of the updated I/O components and their integration with the parsing pipeline.
