+++ src/export/pgn.ts
@@ -1,0 +1,1 @@
+export * from '#pgn/index';
