+++ src/io/__tests__/line-reader.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect } from 'bun:test';
 import { mkdir, readFile, writeFile } from 'node:fs/promises';
 import { join } from 'node:path';
 
-import { openLineStream, readLines } from '#pgn/line-reader';
+import { openLineStream, readLines } from '#io/line-reader';
 
 const FIXTURES_DIR = join(new URL('../../../test/fixtures', import.meta.url).pathname);
 const CRLF_FIXTURE = join(FIXTURES_DIR, 'crlf-endings.pgn');
