+++ src/pgn/index.ts
@@ -1,0 +1,2 @@
+export { parsePGN } from '#pgn/parse-pgn';
+export type { ParsedGame, ParsePgnOptions } from '#types/parse-pgn';
