+++ src/replay/index.ts
@@ -1,0 +1,2 @@
+export type { ReplayMode } from '#replay/replay-policy';
+export { resolveReplayMode, resolveEffectiveReplayMode } from '#replay/replay-policy';
